requirements.txt
antiorm==1.2.1
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
db==0.1.1
db-sqlite3==0.0.1
h11==0.14.0
idna==3.4
outcome==1.2.0
psycopg2==2.9.5
pycparser==2.21
PySocks==1.7.1
python-decouple==3.6
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
wsproto==1.2.0
